requirements.txt
streamlit==1.31.1
google-generativeai==0.4.1